Add files via upload
Finished these points from the google drive:

Add Navtabs for Utility Functions and the subgroups (check the list and the comment that I added below.
Change the name that pops up in the tab when you open the website. It’s currently About Libra. Change it to Libra: fully autuomated machine learning in one-liners.
At the top of modeling queries add the header and the small place for a blurb titled General Dataset Guidelines. This should not be added to the navtab.